+++ src/dipdup/indexes/starknet_node.py
@@ -1,7 +1,8 @@
+from typing import Any
 from typing import Generic
 
 from dipdup.fetcher import BufferT
 from dipdup.indexes.evm_node import EvmNodeFetcher
 
 
-class StarknetNodeFetcher(EvmNodeFetcher, Generic[BufferT]): ...
+class StarknetNodeFetcher(EvmNodeFetcher[Any], Generic[BufferT]): ...
